fix: models book feat: findbuddy

+++ lib/home/models/book_models.dart
@@ -45,7 +45,7 @@ class Fields {
     String isbn10;
     int pageCount;
     String categories;
-    int averageRating;
+    double averageRating;
     int ratingsCount;
     String smallThumbnail;
     String thumbnail;
